Solution to Check Opposite Signs

+++ Check Opposite Signs@@67882a731b4392d8fc1577e1/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    int num1,num2;
-    scanf("%d%d",&num1,&num2);
+    float num1,num2;
+    scanf("%f%f",&num1,&num2);
     float r = num1/num2;
     if(!(r<0)){
         printf("False");
